updated today yo,
We are getting ready to head out in the morning for our trip to Landrum! Can you believe it's
already time? So a lot has happened since my last update.

I had Elder's school last week for starters. It was the most amazing spiritual feast I couldn've
ever imagined. I feel so determined to work hard to build a strong routine and reach out to shepherd
even more. Not that I've done much already anyway. But still, I want to work all the harder at
everything that Jehovah's entrusted to me. It's such a privilege to be able to enjoy that theocratic
education.

Anyway, Metroid Dread came out the Friday after school was out. I had a little bit of overflow work
from taking the week off, but not bad. I'm just about caught all the way up. So I'm hoping to wrap
up the extra today and get the remaining stuff that carried over from before school finished
tomorrow on the road to Jackson, Tennessee.

We'll be caravanning together with the Tulsa crew over to Jackson. Stay the night and then head out
to Landrum. Hoping to get there about 7, or sunset. The temperature should be around the mid-50s if
you can imagine such a thing. So we'll see how that goes, but already my hopes are sky high.

I tried watching a new show today, "MAID" on Netflix... not sure how well I like it but I guess it's
what I'd attribute to the lackluster emoji for today!

Bon apetit! Safe travels. Bon voyage et tous that.